requirements.txt
Flask==2.0.2
Flask_Cors==4.0.0
PyYAML==6.0.1
